refactor preset filters: derive it from filters and place it in the PresetFilters scope
